made themePicker component
