+++ codex-ui/src/vue/components/theme/index.ts
@@ -1,0 +1,3 @@
+import ThemePicker from './ThemePicker.vue';
+
+export { ThemePicker };
